requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
Pillow==10.0.1
django-extensions==3.2.3